add page inert to solve focus vampire
There is a vampire and it steals the focus, something with some kind of highly aggressive `requestAnimationFrame` loop which sucks out all of the focus.

Making the page inert whilst one of the custom windows is open solves it. Is it nice? No, I'd rather not have this but I can't see another workaround for now.